GuiCommon: Fixed HDR colours tree item properties.
